Group existing tests into smaller test files
