+++ src/integration-test/tests/citizen/defence/defence_handoff_test.ts
@@ -1,0 +1,53 @@
+import { PartyType } from 'integration-test/data/party-type'
+import { createClaimData } from 'integration-test/data/test-data'
+import { Helper } from 'integration-test/tests/citizen/endToEnd/steps/helper'
+import I = CodeceptJS.I
+import { DefenceType } from 'integration-test/data/defence-type'
+
+const helperSteps: Helper = new Helper()
+
+Feature('Respond to claim: handoff journey')
+
+Scenario('I can see send your response by email page when I admit part of the claim @citizen', function* (I: I) {
+  const claimantEmail: string = yield I.createCitizenUser()
+  const defendantEmail: string = yield I.createCitizenUser()
+
+  const claimData: ClaimData = createClaimData(PartyType.INDIVIDUAL, PartyType.INDIVIDUAL)
+  const defendant: Party = claimData.defendants[0]
+  const claimant: Party = claimData.claimants[0]
+
+  const claimRef: string = yield I.createClaim(claimData, claimantEmail)
+
+  yield helperSteps.enterPinNumber(claimRef, claimantEmail)
+  helperSteps.finishResponseWithHandOff(claimRef, defendant, claimant, defendantEmail, DefenceType.PART_ADMISSION)
+})
+
+Scenario('I can see send your response by email page when I reject all of the claim with counter claim @citizen', function* (I: I) {
+  const claimantEmail: string = yield I.createCitizenUser()
+  const defendantEmail: string = yield I.createCitizenUser()
+
+  const claimData: ClaimData = createClaimData(PartyType.INDIVIDUAL, PartyType.INDIVIDUAL)
+  const defendant: Party = claimData.defendants[0]
+  const claimant: Party = claimData.claimants[0]
+
+  const claimRef: string = yield I.createClaim(claimData, claimantEmail)
+
+  yield helperSteps.enterPinNumber(claimRef, claimantEmail)
+  helperSteps.finishResponseWithHandOff(claimRef, defendant, claimant, defendantEmail, DefenceType.FULL_REJECTION_WITH_COUNTER_CLAIM)
+})
+
+Scenario('I can see send your response by email page when I reject all of the claim with amount paid less than claimed amount @citizen', function* (I: I) {
+  const claimantEmail: string = yield I.createCitizenUser()
+  const defendantEmail: string = yield I.createCitizenUser()
+
+  const claimData: ClaimData = createClaimData(PartyType.INDIVIDUAL, PartyType.INDIVIDUAL)
+  const defendant: Party = claimData.defendants[0]
+  const claimant: Party = claimData.claimants[0]
+
+  const claimRef: string = yield I.createClaim(claimData, claimantEmail)
+
+  yield helperSteps.enterPinNumber(claimRef, claimantEmail)
+  helperSteps.finishResponseWithHandOff(claimRef, defendant, claimant,
+    defendantEmail, DefenceType.FULL_REJECTION_BECAUSE_ALREADY_PAID_LESS_THAN_CLAIMED_AMOUNT
+  )
+})
